Move flash mint utils
