+++ src/hooks/useFlashMintQuote.ts
@@ -8,9 +8,9 @@ import { Token } from 'constants/tokens'
 import { useNetwork } from 'hooks/useNetwork'
 import { useBalanceData } from 'providers/Balances'
 import { GasStation } from 'utils/api/gasStation'
+import { getConfiguredZeroExApi } from 'utils/api/zeroExApi'
 import { getNetworkKey } from 'utils/api/zeroExUtils'
 import { getAddressForToken } from 'utils/tokens'
-import { getConfiguredZeroExApi } from 'utils/zeroExApi'
 
 import { useIssuanceQuote } from './issuance/useIssuanceQuote'
 import { getEnhancedFlashMintLeveragedQuote } from './useBestQuote/flashMintLeveraged'
